new correlation scheme and smoke test
Key Updates

Smart-incubator now supports continuous [-1,1] correlation: per-cycle behaviour is randomly selected as paired/no-US/random with correct probabilities and logged (Smart_incubator/Firmware/run_experiment_cycle.py (lines 36-394)). OLED output, experiment params, and logging metadata all handle floats cleanly (Smart_incubator/Firmware/main.py (lines 29-270), Smart_incubator/Firmware/main_with_server.py (lines 41-345), Smart_incubator/Firmware/oled_display.py (lines 30-63), Smart_incubator/Firmware/sd_logger.py (lines 65-95)).
Web UI accepts numeric correlation with clamping, sync script can override float values, and legacy discrete modes were removed (Smart_incubator/Firmware/web_server.py (lines 15-236), Smart_incubator/sync_firmware.py (lines 40-588)).
Added host-side dry smoke test that stubs hardware modules, accelerates time, and runs the real firmware logic for a set number of cycles without ESP32 hardware (Smart_incubator/tests/dry_smoke_test.py (lines 1-257)).

+++ Smart_incubator/Firmware/main.py
@@ -33,7 +33,7 @@ min_interval = 200      # minutes (not seconds!)
 max_interval = 400      # minutes (not seconds!)
 us_duration_seconds = 1800     # US window length in seconds (30 minutes)
 heat_duration_seconds = 1800     # Heat shock duration in seconds (30 minutes)
-correlation = 1       # US precedes heat shock at end of cycle
+correlation = 1.0     # Range [-1, 1]: -1 = no US, 0 = random, 1 = paired US before heat
 
 def init_output_pins():
     """Initialize all output pins with PWM."""
@@ -160,7 +160,7 @@ def init_system():
             'heat_duration': heat_duration_minutes,
             'us_duration_seconds': int(us_duration_seconds),
             'heat_duration_seconds': int(heat_duration_seconds),
-            'correlation': int(correlation)
+            'correlation': float(correlation)
         }
         experiment_logger = ExperimentLogger(experiment_params)
